Created indexLinkingToSamePage.html and aboutMeLinkingToSamePage.html to showcase the concept of linking to same page section
In this example, we used the atributes 'id's placed on each <div> tag to create an unordered list.
each item text will be nested within an anchor tag <a> that contains the id of each <div>. Example:
<ul>
	<li><a href="#introduction">Introduction</a></li>
	<li><a href="#habitat">Habitat</a></li>
	<li><a href="media">Media</a></li>
</ul>

This is what is was added to allow users to scroll automatically to a specific part of the site, depending on
where the 'id' atributes were placed within each <div> tag.